refactor: rename daemon coding endpoint
